Changing to typeScript
